Read a network location from the coordinate module
_location_from_coords lived in obs.py yet served model and comparison
too, and it reads the same coordinates NodeCoords and ReachCoords write.
It moves beside them as network_location. _at_from_coords stays in
obs.py, since coercing to NodeObservation's `at` types is its own affair.
