+++ src/modelskill/comparison/_comparison.py
@@ -31,10 +31,13 @@ from ..obs import (
     TrackObservation,
     NodeObservation,
     ReachObservation,
-    _location_from_coords,
 )
 from ..model import PointModelResult, TrackModelResult, VerticalModelResult
-from ..timeseries._coords import NETWORK_LOCATION_COORDS, _coordinate_values
+from ..timeseries._coords import (
+    NETWORK_LOCATION_COORDS,
+    _coordinate_values,
+    network_location,
+)
 from ..timeseries._timeseries import (
     _normalize_time_to_ns,
     _validate_data_var_name,
@@ -667,7 +670,7 @@ class Comparer:
     @property
     def _at(self) -> Any:
         """Where this comparer sits, in the form NodeObservation.at takes"""
-        return _location_from_coords(self.data)
+        return network_location(self.data)
 
     def _coordinate_values(self, coord: str) -> Any:
         """Get coordinate values if they exist, otherwise return None"""
